trying to fix the ints

+++ src/libxtp/threecenter_gwbse.cc
@@ -204,11 +204,11 @@ void TCMatrix_gwbse::MatrixTensorMatrixProduct(
     std::vector<Eigen::MatrixXd>& block,
     const std::vector<Eigen::MatrixXd>& symmstorage,
     const Eigen::MatrixXd& dftn, const Eigen::MatrixXd& dftm) const {
-  for (long int k = 0; k < symmstorage.size(); ++k) {
+  for (unsigned k = 0; k < symmstorage.size(); ++k) {
     const Eigen::MatrixXd& matrix = symmstorage[k];
     Eigen::MatrixXd threec_inMo =
         dftn.transpose() * matrix.selfadjointView<Eigen::Lower>() * dftm;
-    for (long int i = 0; i < threec_inMo.cols(); ++i) {
+    for (int i = 0; i < threec_inMo.cols(); ++i) {
       block[i].col(k) = threec_inMo.col(i);
     }
   }
